meraki_dashboard_api: Remove params support
It relied on overriding an internal requests library function.
To switch to httpx (to avoid introducing a new dependency),
this would have to be implemented differently
(or would not be possible).

Remove params support for now since it is not used in nac-collector,
rather than leaving it with potentially unexpected behavior
compared to the original dashboard-api-python.